Bump angular versions to 14 Refactor the project structure to comply recommendations
