+++ .gitignore
@@ -1,15 +1,42 @@
-node_modules
-src/node_modules
+# See http://help.github.com/ignore-files/ for more about ignoring files.
+
+# Compiled output
+/dist
+/tmp
+/out-tsc
+/bazel-out
+
+# Node
+/node_modules
 npm-debug.log
-coverage
-dist
-build
-typings
-src/*.js
-src/*.map
-src/*.d.ts
-src/*.ngfactory.ts
-angular2-toaster.d.ts
-demo/webpack/bundle.js
-demo/webpack/bundle.js.map
+yarn-error.log
+
+# IDEs and editors
+.idea/
+.project
+.classpath
+.c9/
+*.launch
+.settings/
+*.sublime-workspace
+
+# Visual Studio Code
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+.history/*
+
+# Miscellaneous
+/.angular/cache
+.sass-cache/
+/connect.lock
+/coverage
+/libpeerconnection.log
+testem.log
+/typings
+
+# System files
 .DS_Store
+Thumbs.db
