+++ .travis.yml
@@ -8,7 +8,7 @@ before_install:
   - export CHROME_BIN=chromium-browser
   - export DISPLAY=:99.0
   - sh -e /etc/init.d/xvfb start
-install: 
+install:
   - npm install
 script:
   - ng test angular2-toaster --watch=false --code-coverage
