+++ package.json
@@ -5,52 +5,27 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "build-prod": "ng build --prod && npm run sass && npm run uglifycss && npm run copy",
-    "sass": "sass src/angular2-toaster/src/lib/toaster.scss src/angular2-toaster/src/lib/toaster.css",
-    "uglifycss": "node_modules/.bin/uglifycss src/angular2-toaster/src/lib/toaster.css > src/angular2-toaster/src/lib/toaster.min.css",
-    "copy": "cpx LICENSE dist/angular2-toaster && cpx README.md dist/angular2-toaster && npm run copystyles",
-    "copystyles": "cpx \"src/angular2-toaster/src/lib/*.{css,css.map,min.css,scss}\" dist/angular2-toaster"
+    "watch": "ng build --watch --configuration development"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.0.0",
-    "@angular/common": "~11.0.0",
-    "@angular/compiler": "~11.0.0",
-    "@angular/core": "~11.0.0",
-    "@angular/forms": "~11.0.0",
-    "@angular/platform-browser": "~11.0.0",
-    "@angular/platform-browser-dynamic": "~11.0.0",
-    "@angular/router": "~11.0.0",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "@angular/animations": "^14.2.0",
+    "@angular/common": "^14.2.0",
+    "@angular/compiler": "^14.2.0",
+    "@angular/core": "^14.2.0",
+    "@angular/forms": "^14.2.0",
+    "@angular/platform-browser": "^14.2.0",
+    "@angular/platform-browser-dynamic": "^14.2.0",
+    "@angular/router": "^14.2.0",
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.1",
-    "@angular/cli": "~11.0.1",
-    "@angular/compiler-cli": "~11.0.0",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "coveralls": "3.1.0",
-    "cpx": "^1.5.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.1",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^11.0.2",
-    "protractor": "~7.0.0",
-    "sass": "1.29.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.2",
-    "uglifycss": "0.0.29"
+    "@angular-devkit/build-angular": "^14.2.5",
+    "@angular/cli": "~14.2.5",
+    "@angular/compiler-cli": "^14.2.0",
+    "ng-packagr": "^14.2.0",
+    "typescript": "~4.7.2"
   }
 }
