+++ projects/angular2-toaster/ng-package.json
@@ -1,6 +1,9 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
   "dest": "../../dist/angular2-toaster",
+  "assets": [
+    "./**/lib/toaster.scss"
+  ],
   "lib": {
     "entryFile": "src/public-api.ts"
   }
