+++ projects/angular2-toaster/package.json
@@ -1,15 +1,15 @@
 {
-  "name": "angular2-toaster",
+  "name": "@dkachko/angular2-toaster",
   "description": "An Angular Toaster Notification library based on AngularJS-Toaster",
-  "version": "11.0.1",
+  "version": "14.0.1",
   "author": "Stabzs",
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/stabzs/Angular2-Toaster.git"
+    "url": "https://github.com/dkachko/Angular2-Toaster.git"
   },
   "bugs": {
-    "url": "https://github.com/stabzs/Angular2-Toaster/issues"
+    "url": "https://github.com/dkachko/Angular2-Toaster/issues"
   },
   "keywords": [
     "angular",
@@ -24,13 +24,13 @@
     "growl"
   ],
   "peerDependencies": {
-    "@angular/common": "^11.0.0",
-    "@angular/compiler": "^11.0.0",
-    "@angular/core": "^11.0.0",
-    "rxjs": "^6.5.3"
+    "@angular/common": "^14.2.0",
+    "@angular/compiler": "^14.2.0",
+    "@angular/core": "^14.2.0",
+    "rxjs": "^7.5.0"
   },
   "dependencies": {
-    "tslib": "^2.0.0"
+    "tslib": "^2.4.0"
   },
   "sideEffects": false
 }
