+++ projects/angular2-toaster/src/lib/bodyOutputType.ts
