+++ projects/angular2-toaster/src/lib/clearWrapper.ts
