+++ projects/angular2-toaster/src/lib/toast.component.spec.ts
