+++ projects/angular2-toaster/src/lib/toast.ts
