+++ projects/angular2-toaster/src/lib/toaster-container.component.spec.ts
