+++ projects/angular2-toaster/src/lib/toaster.scss
