+++ projects/angular2-toaster/src/lib/transitions.ts
