+++ projects/angular2-toaster/src/lib/trust-html.pipe.ts
