+++ projects/angular2-toaster/src/public-api.ts
@@ -7,5 +7,6 @@ export * from './lib/toaster.service';
 export * from './lib/clearWrapper';
 export * from './lib/toaster-config';
 export * from './lib/toast';
+export * from './lib/toast.component';
 export * from './lib/bodyOutputType';
 export * from './lib/toaster.module';
