+++ projects/angular2-toaster/tsconfig.lib.json
