+++ projects/demo/karma.conf.js
@@ -3,7 +3,7 @@
 
 module.exports = function (config) {
   config.set({
-    basePath: '',
+    basePath: '../../temp/demo',
     frameworks: ['jasmine', '@angular-devkit/build-angular'],
     plugins: [
       require('karma-jasmine'),
@@ -16,15 +16,9 @@ module.exports = function (config) {
       clearContext: false // leave Jasmine Spec Runner output visible in browser
     },
     coverageIstanbulReporter: {
-      dir: require('path').join(__dirname, '../../coverage/angular2-toaster'),
+      dir: require('path').join(__dirname, '../../coverage/demo'),
       reports: ['html', 'lcovonly', 'text-summary'],
       fixWebpackSourcePaths: true
-    },
-    customLaunchers: {
-      ChromeCI: {
-        base: `${process.env['TRAVIS'] ? 'ChromeHeadless' : 'Chrome'}`,
-        flags: process.env['TRAVIS'] ? ['--no-sandbox'] : [],
-      },
     },
     reporters: ['progress', 'kjhtml'],
     port: 9876,
