+++ projects/demo/src/styles.scss
@@ -1,2 +1,2 @@
 /* You can add global styles to this file, and also import other style files */
-@import "../../angular2-toaster/src/lib/toaster.scss"
+@import "../../angular2-toaster/src/lib/toaster.css";
